requirements.txt
aiofile==3.8.8
aiofiles==23.2.1
aiohttp==3.9.3
aiomysql==0.1.1
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Authlib==1.3.0
autoflake==2.3.1
bcrypt==4.0.1
billiard==4.2.0
black==24.3.0
caio==0.9.13
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
dynaconf==3.2.5
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.2
fastapi-cache2==0.2.1
fastapi-pagination==0.12.19
fastapi-sa==0.2.0
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
flake8==7.0.0
flower==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
humanize==4.9.0
idna==3.6
imagekitio==4.0.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
minio==7.2.5
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.15
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyflakes==3.2.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyright==1.1.360
pytest==8.1.1
pytest-asyncio==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
sentry-sdk==1.44.1
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
sqladmin==0.16.1
SQLAlchemy==2.0.29
starlette==0.37.2
time-machine==2.14.0
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.21.1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
WTForms==3.1.2
yarl==1.9.4
